Programming Task Outlines
Need to get the lessons intended in each week to at least be outlined with a phrase. Need to assign the right source code blocks for folks to dissect and actually attach it to the lessons themselves.

To do: Consider how to use the intersecting weeks to the fullest extent.